requirements.txt
Django>=3.0
django_taggit>=1.2.0
